simplify directory structure
